Demonstrate order validation in CI
Many of the init steps are order-specific, depending on a series of mutations or additions to the form before it is ready for use. These steps are likely interchangeable, as their order is different in offline mode than online:

- `_swapTheme`
- `_addBranding`

This change will cause the order of the performed init steps in the test to be incorrect.